Move extensions into utils/extension
