+++ app/src/main/kotlin/com/absinthe/libchecker/ui/fragment/detail/AppBundleBottomSheetDialogFragment.kt
@@ -3,10 +3,10 @@ package com.absinthe.libchecker.ui.fragment.detail
 import android.view.ViewGroup
 import com.absinthe.libchecker.R
 import com.absinthe.libchecker.bean.AppBundleItemBean
-import com.absinthe.libchecker.extensions.dp
 import com.absinthe.libchecker.ui.detail.EXTRA_PACKAGE_NAME
 import com.absinthe.libchecker.ui.fragment.BaseBottomSheetViewDialogFragment
 import com.absinthe.libchecker.utils.PackageUtils
+import com.absinthe.libchecker.utils.extensions.dp
 import com.absinthe.libchecker.view.app.BottomSheetHeaderView
 import com.absinthe.libchecker.view.detail.AppBundleBottomSheetView
 import com.absinthe.libchecker.view.detail.AppBundleItemView
