+++ app/src/main/kotlin/com/absinthe/libchecker/ui/fragment/settings/CloudRulesDialogFragment.kt
@@ -9,11 +9,11 @@ import com.absinthe.libchecker.api.request.CloudRuleBundleRequest
 import com.absinthe.libchecker.constant.Constants
 import com.absinthe.libchecker.constant.GlobalValues
 import com.absinthe.libchecker.database.RuleDatabase
-import com.absinthe.libchecker.extensions.addPaddingTop
-import com.absinthe.libchecker.extensions.dp
 import com.absinthe.libchecker.ui.fragment.BaseBottomSheetViewDialogFragment
-import com.absinthe.libchecker.utils.showToast
 import com.absinthe.libchecker.utils.DownloadUtils
+import com.absinthe.libchecker.utils.extensions.addPaddingTop
+import com.absinthe.libchecker.utils.extensions.dp
+import com.absinthe.libchecker.utils.showToast
 import com.absinthe.libchecker.view.app.BottomSheetHeaderView
 import com.absinthe.libchecker.view.settings.CloudRulesDialogView
 import kotlinx.coroutines.Dispatchers
