+++ app/src/main/kotlin/com/absinthe/libchecker/ui/fragment/snapshot/TimeNodeBottomSheetDialogFragment.kt
@@ -3,8 +3,8 @@ package com.absinthe.libchecker.ui.fragment.snapshot
 import android.view.ViewGroup
 import com.absinthe.libchecker.R
 import com.absinthe.libchecker.database.entity.TimeStampItem
-import com.absinthe.libchecker.extensions.dp
 import com.absinthe.libchecker.ui.fragment.BaseBottomSheetViewDialogFragment
+import com.absinthe.libchecker.utils.extensions.dp
 import com.absinthe.libchecker.utils.putArguments
 import com.absinthe.libchecker.view.app.BottomSheetHeaderView
 import com.absinthe.libchecker.view.snapshot.TimeNodeBottomSheetView
