+++ app/src/main/kotlin/com/absinthe/libchecker/ui/fragment/statistics/ClassifyBottomSheetDialogFragment.kt
@@ -7,8 +7,8 @@ import android.view.View
 import android.view.ViewGroup
 import androidx.lifecycle.lifecycleScope
 import com.absinthe.libchecker.database.entity.LCItem
-import com.absinthe.libchecker.extensions.dp
 import com.absinthe.libchecker.ui.fragment.BaseBottomSheetViewDialogFragment
+import com.absinthe.libchecker.utils.extensions.dp
 import com.absinthe.libchecker.view.app.BottomSheetHeaderView
 import com.absinthe.libchecker.view.statistics.ClassifyDialogView
