+++ app/src/main/kotlin/com/absinthe/libchecker/utils/extensions/ContextExtensions.kt
@@ -1,4 +1,4 @@
-package com.absinthe.libchecker.extensions
+package com.absinthe.libchecker.utils.extensions
 
 import android.annotation.SuppressLint
 import android.app.Activity
