+++ app/src/main/kotlin/com/absinthe/libchecker/utils/extensions/ViewPagerKtx.kt
@@ -1,4 +1,4 @@
-package com.absinthe.libchecker.extensions
+package com.absinthe.libchecker.utils.extensions
 
 import android.animation.Animator
 import android.animation.TimeInterpolator
