+++ app/src/main/kotlin/com/absinthe/libchecker/view/snapshot/SnapshotDetailDeletedView.kt
@@ -7,7 +7,7 @@ import android.widget.ImageView
 import androidx.appcompat.widget.AppCompatTextView
 import androidx.core.view.marginTop
 import com.absinthe.libchecker.R
-import com.absinthe.libchecker.extensions.getResourceIdByAttr
+import com.absinthe.libchecker.utils.extensions.getResourceIdByAttr
 import com.absinthe.libchecker.view.AViewGroup
 
 class SnapshotDetailDeletedView(context: Context) : AViewGroup(context) {
