+++ app/src/main/kotlin/com/absinthe/libchecker/view/snapshot/SnapshotDetailNewInstallView.kt
@@ -7,7 +7,7 @@ import android.widget.ImageView
 import androidx.appcompat.widget.AppCompatTextView
 import androidx.core.view.marginTop
 import com.absinthe.libchecker.R
-import com.absinthe.libchecker.extensions.getResourceIdByAttr
+import com.absinthe.libchecker.utils.extensions.getResourceIdByAttr
 import com.absinthe.libchecker.view.AViewGroup
 
 class SnapshotDetailNewInstallView(context: Context) : AViewGroup(context) {
